search

+++ app/Http/Controllers/API/UserController.php
@@ -17,6 +17,7 @@ class UserController extends Controller
     public function __construct()
     {
         $this->middleware('auth:api');
+        // $this->authorize('isAdmin');
     }
     /**
      * Display a listing of the resource.
@@ -25,7 +26,14 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        // $this->authorize('isAdmin');
+        if(\Gate::allows('isAdmin')) {
+            return User::latest()->paginate(10);
+        }
+        if(\Gate::allows('isManeger')) {
+            $user = auth('api')->user();
+            return User::where('franchise', $user->franchise)->latest()->paginate(10);
+        }
     }
 
     /**
@@ -72,10 +80,28 @@ class UserController extends Controller
     public function updateProfile(Request $request)
     {
         $user = auth('api')->user();
-        if($request->photo) {
+        $this->validate($request, [
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,'.$user->id,
+            'password' => 'sometimes|required|string|min:6'
+        ]);
+        if(!empty($request->password)) {
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+
+        $currentPhoto = $user->photo;
+        if($request->photo != $currentPhoto) {
             $name = time().'.' . explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
             \Image::make($request->photo)->save(public_path('images/profile/').$name);
-        }        
+            $request->merge(['photo' => $name]);
+            $user->photo = $request->photo;
+            $userPhoto = public_path('images/profile/').$currentPhoto;
+            if(file_exists($userPhoto)) {
+                @unlink($userPhoto);
+            }
+        }
+        $user->update($request->all());
+        return ['message', 'User updated successfully' . $user->password];
     }
 
     /**
@@ -96,7 +122,9 @@ class UserController extends Controller
             'type' => 'required',
             'franchise' => 'required'
         ]);
-
+        if(!empty($request->password)) {
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }        
         $user->update($request->all());
         return ['message' => 'User updated', $user];
     }
@@ -109,9 +137,31 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
 
         return ['message' => 'User Deleted'];
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function search()
+    {
+        if ($search = \Request::get('q')) {
+            $users = User::where(function($query) use ($search) {
+                $query->where('name', 'LIKE', "%$search%")
+                    ->orWhere('email','LIKE',"%$search%")
+                    ->orWhere('type','LIKE',"%$search%")
+                    ->orWhere('franchise','LIKE',"%$search%");
+            })->paginate(20);
+        } else {
+            $users = User::latest()->paginate(10);
+        }
+        return $users;
     }
 }
